+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 - Default decoding structure of `Parser` is object. (You won't notice that unless you use `Parser` class directly)
+- `Items::__construct` accepts the options array instead of separate arguments. (You won't notice that unless you instantiate `Items` class directly)
 
 <br>
 <br>